feat: regroup the prompts in a shared helper
